+++ src/modules/modules.index.ts
@@ -1,6 +1,5 @@
-import { AngularBuildModule } from "./angular-build/angular-build.module";
-import { Kbs6LibModule } from "./kbs6-lib/kbs6-lib.module";
-import { SettingsModule } from "./settings/settings.module";
+import { DynamicSamplesModule } from "./dynamic-samples/dynamic-samples.module";
+import { SamplesModule } from "./samples/samples.module";
 
 /**
  * Represents a utility class for managing modules.
@@ -15,14 +14,12 @@ export class Modules {
      */
     public static setModules(): void {
         // Get modules instances
-        const kbs6LibModule = new Kbs6LibModule();
-        const angularBuildModule = new AngularBuildModule();
-        const settingsModule = new SettingsModule();
+        const samplesModule = new SamplesModule();
+        const dynamicSamplesModule = new DynamicSamplesModule();
 
         // Add module instances to the modules object
-        Modules.modules[kbs6LibModule.getId()] = kbs6LibModule;
-        Modules.modules[angularBuildModule.getId()] = angularBuildModule;
-        Modules.modules[settingsModule.getId()] = settingsModule;
+        Modules.modules[samplesModule.getId()] = samplesModule;
+        Modules.modules[dynamicSamplesModule.getId()] = dynamicSamplesModule;
     }
 
     /**