+++ src/store.ts
@@ -17,34 +17,6 @@ export class Store {
     public static panelViewWebview: vscode.Webview | undefined;
     public static panelViewWebviewContext: vscode.ExtensionContext | undefined;
 
-    public static icons = [
-        `arrows-right-left`,
-        `book`,
-        `building-castle`,
-        `check-circle`,
-        `chevron-left`,
-        `chevron-right`,
-        `copy`,
-        `download-cloud`,
-        `edit`,
-        `external-link`,
-        `eye`,
-        `hammer`,
-        `loader`,
-        `mountain`,
-        `package`,
-        `palette`,
-        `plane-departure`,
-        `plane`,
-        `play`,
-        `search`,
-        `settings`,
-        `square`,
-        `trees`,
-        `upload-cloud`,
-        `x-circle`
-    ];
-
     public static terminalColorsCssVars = [
         `--vscode-terminal-ansiBlack`,
         `--vscode-terminal-ansiRed`,